chore: update README.md installation instructions for yarn
The installation instructions for yarn in the README.md file have been updated to fix an issue with the installation command. Previously, the command `yarn global add file:.` was used, which did not work correctly. The new command `yarn global add file:$PWD` is now used to correctly install the plugin. This change ensures that users can successfully install the plugin using yarn.